unification: unify within objects, reuse keys. nearly
Not quite right still.

Need a smarter Map.merge function to that manages to set keys
with non-union values to optional decoder.

Need to make sure we aren't shooting ourselves in the foot for future
recursive detection.

Potentially swap between combine-objects and unify-decoders at an object
level depending on how similar any given objects in an array [ or
globally? ] are.

ie. [ obja | objb | objb | obja ]

may be preferable to

{ a: optional, b : optional, c : optional, d : optional }

if the objects don't really have much in common. Seems to become a
search problem.